Ajuste fino, projeto sem erros!

+++ src/main/java/com/doback/E_rank/controller/AmizadesController.java
@@ -1,7 +1,6 @@
 package com.doback.E_rank.controller;
 
 import com.doback.E_rank.entity.Amizades;
-import com.doback.E_rank.entity.Times;
 import com.doback.E_rank.facade.AmizadesFacade;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -24,7 +23,7 @@ public class AmizadesController {
     }
 
     @GetMapping("/{id}")
-    public Amizades obterAmizade(@PathVariable Long id) {
+    public Amizades obterAmizade(@PathVariable int id) {
         return amizadeFacade.buscarAmizadePorId(id);
     }
 
@@ -36,13 +35,13 @@ public class AmizadesController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void excluirAmizade(@PathVariable Long id) {
+    public void excluirAmizade(@PathVariable int id) {
         amizadeFacade.excluirAmizade(id);
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public void atualizarAmizades(@PathVariable long id,@RequestBody Amizades amizades){
+    public void atualizarAmizades(@PathVariable int id,@RequestBody Amizades amizades){
         amizadeFacade.atualizarAmizades(id, amizades);
     }
